Add files via upload
I seem to be only be able to connect one transformer to my power network? One on the "input" and one "output" side. My setup is as follows, I have a main network consisting of large powerpoles and than I want to use transformers to make brances of the network. However, I am only able to connect one transformer, the next one does not seem to "transmit" any power?